+++ src/main/java/com/example/DongNaeJoGak/domain/auth/annotation/AuthenticatedMemberResolver.java
@@ -1,6 +1,6 @@
 package com.example.DongNaeJoGak.domain.auth.annotation;
 
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 import com.example.DongNaeJoGak.global.apiPayload.code.status.error.TokenErrorStatus;
 import com.example.DongNaeJoGak.global.apiPayload.exception.TokenException;
 import jakarta.servlet.http.HttpServletRequest;
