+++ src/main/java/com/example/DongNaeJoGak/domain/auth/dto/request/OAuthRequestDTO.java
@@ -1,6 +1,6 @@
 package com.example.DongNaeJoGak.domain.auth.dto.request;
 
-import com.example.DongNaeJoGak.domain.member.entity.enums.ProviderType;
+import com.example.DongNaeJoGak.domain.idea.member.entity.enums.ProviderType;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
