+++ src/main/java/com/example/DongNaeJoGak/domain/auth/service/AbstractOAuth2Service.java
@@ -2,9 +2,9 @@ package com.example.DongNaeJoGak.domain.auth.service;
 
 import com.example.DongNaeJoGak.domain.auth.dto.request.OAuthRequestDTO;
 import com.example.DongNaeJoGak.domain.auth.dto.response.OAuthResponseDTO;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
-import com.example.DongNaeJoGak.domain.member.entity.enums.ProviderType;
-import com.example.DongNaeJoGak.domain.member.repository.MemberRepository;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.enums.ProviderType;
+import com.example.DongNaeJoGak.domain.idea.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 
 import java.util.Optional;
