+++ src/main/java/com/example/DongNaeJoGak/domain/auth/service/OAuth2Service.java
@@ -1,7 +1,7 @@
 package com.example.DongNaeJoGak.domain.auth.service;
 
 import com.example.DongNaeJoGak.domain.auth.dto.response.OAuthResponseDTO;
-import com.example.DongNaeJoGak.domain.member.entity.enums.ProviderType;
+import com.example.DongNaeJoGak.domain.idea.member.entity.enums.ProviderType;
 
 public interface OAuth2Service {
