+++ src/main/java/com/example/DongNaeJoGak/domain/auth/service/TokenService.java
@@ -1,7 +1,7 @@
 package com.example.DongNaeJoGak.domain.auth.service;
 
 import com.example.DongNaeJoGak.domain.auth.dto.response.OAuthResponseDTO;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 
 public interface TokenService {
     OAuthResponseDTO.LoginResponse createLoginToken(Member member);
