+++ src/main/java/com/example/DongNaeJoGak/domain/auth/service/TokenServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.DongNaeJoGak.domain.auth.service;
 
 import com.example.DongNaeJoGak.domain.auth.dto.response.OAuthResponseDTO;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 import com.example.DongNaeJoGak.global.security.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
