+++ src/main/java/com/example/DongNaeJoGak/domain/comment/entity/CommentReport.java
@@ -1,7 +1,7 @@
 package com.example.DongNaeJoGak.domain.comment.entity;
 
 import com.example.DongNaeJoGak.domain.comment.entity.enums.CommentReportType;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.*;
