+++ src/main/java/com/example/DongNaeJoGak/domain/comment/service/CommentReportServiceImpl.java
@@ -6,8 +6,8 @@ import com.example.DongNaeJoGak.domain.comment.entity.Comment;
 import com.example.DongNaeJoGak.domain.comment.entity.CommentReport;
 import com.example.DongNaeJoGak.domain.comment.repository.CommentReportRepository;
 import com.example.DongNaeJoGak.domain.comment.repository.CommentRepository;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
-import com.example.DongNaeJoGak.domain.member.repository.MemberRepository;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.repository.MemberRepository;
 import com.example.DongNaeJoGak.global.apiPayload.code.status.error.CommentErrorStatus;
 import com.example.DongNaeJoGak.global.apiPayload.code.status.error.MemberErrorStatus;
 import com.example.DongNaeJoGak.global.apiPayload.exception.CommentException;
