+++ src/main/java/com/example/DongNaeJoGak/domain/comment/service/CommentServiceImpl.java
@@ -7,8 +7,8 @@ import com.example.DongNaeJoGak.domain.comment.entity.Comment;
 import com.example.DongNaeJoGak.domain.comment.repository.CommentRepository;
 import com.example.DongNaeJoGak.domain.idea.entity.Idea;
 import com.example.DongNaeJoGak.domain.idea.repository.IdeaRepository;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
-import com.example.DongNaeJoGak.domain.member.repository.MemberRepository;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.repository.MemberRepository;
 import com.example.DongNaeJoGak.global.apiPayload.exception.CommentException;
 import com.example.DongNaeJoGak.global.apiPayload.exception.IdeaException;
 import com.example.DongNaeJoGak.global.apiPayload.exception.MemberException;
