+++ src/main/java/com/example/DongNaeJoGak/domain/idea/entity/Idea.java
@@ -2,7 +2,7 @@ package com.example.DongNaeJoGak.domain.idea.entity;
 
 import com.example.DongNaeJoGak.domain.comment.entity.Comment;
 import com.example.DongNaeJoGak.domain.idea.entity.enums.IdeaStatus;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 import com.example.DongNaeJoGak.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
