+++ src/main/java/com/example/DongNaeJoGak/domain/idea/member/entity/Member.java
@@ -1,8 +1,8 @@
-package com.example.DongNaeJoGak.domain.member.entity;
+package com.example.DongNaeJoGak.domain.idea.member.entity;
 
 import com.example.DongNaeJoGak.domain.comment.entity.Comment;
 import com.example.DongNaeJoGak.domain.idea.entity.Idea;
-import com.example.DongNaeJoGak.domain.member.entity.enums.ProviderType;
+import com.example.DongNaeJoGak.domain.idea.member.entity.enums.ProviderType;
 import jakarta.persistence.*;
 import lombok.*;
