+++ src/main/java/com/example/DongNaeJoGak/domain/idea/member/repository/MemberRepository.java
@@ -1,8 +1,8 @@
-package com.example.DongNaeJoGak.domain.member.repository;
+package com.example.DongNaeJoGak.domain.idea.member.repository;
 
 
-import com.example.DongNaeJoGak.domain.member.entity.Member;
-import com.example.DongNaeJoGak.domain.member.entity.enums.ProviderType;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.enums.ProviderType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
