+++ src/main/java/com/example/DongNaeJoGak/domain/idea/repository/IdeaReactionRepository.java
@@ -2,7 +2,7 @@ package com.example.DongNaeJoGak.domain.idea.repository;
 
 import com.example.DongNaeJoGak.domain.idea.entity.Idea;
 import com.example.DongNaeJoGak.domain.idea.entity.IdeaReaction;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IdeaReactionRepository extends JpaRepository<IdeaReaction, Long> {
