+++ src/main/java/com/example/DongNaeJoGak/domain/idea/service/IdeaReactionService.java
@@ -2,7 +2,7 @@ package com.example.DongNaeJoGak.domain.idea.service;
 
 import com.example.DongNaeJoGak.domain.idea.dto.response.IdeaReactionResponseDTO;
 import com.example.DongNaeJoGak.domain.idea.entity.enums.IdeaReactionType;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
+import com.example.DongNaeJoGak.domain.idea.member.entity.Member;
 
 public interface IdeaReactionService {
