+++ src/main/java/com/example/DongNaeJoGak/domain/idea/service/IdeaService.java
@@ -2,8 +2,6 @@ package com.example.DongNaeJoGak.domain.idea.service;
 
 import com.example.DongNaeJoGak.domain.idea.dto.request.IdeaRequestDTO;
 import com.example.DongNaeJoGak.domain.idea.dto.response.IdeaResponseDTO;
-import com.example.DongNaeJoGak.domain.idea.entity.enums.IdeaReactionType;
-import com.example.DongNaeJoGak.domain.member.entity.Member;
 
 public interface IdeaService {
