--- src/main/java/com/example/DongNaeJoGak/domain/member/entity/enums/ProviderType.java
@@ -1,5 +1,0 @@
-package com.example.DongNaeJoGak.domain.member.entity.enums;
-
-public enum ProviderType {
-    NAVER, GOOGLE
-}
