+++ src/main/resources/application.yml
@@ -23,7 +23,6 @@ spring:
           client-id: ${NAVER_CLIENT_ID}
           client-secret: ${NAVER_CLIENT_SECRET}
           client-name: Naver
-          redirect-uri: http://localhost:3000/login/oauth2/code/naver
           authorization-grant-type: authorization_code
           authorization-uri: https://nid.naver.com/oauth2.0/authorize
           token-uri: https://nid.naver.com/oauth2.0/token
